+++ directdm/num/num_input.py
@@ -6,11 +6,13 @@ import warnings
 from directdm.run import rge
 
 
-#---------------------------------------------------------------------#
-#                                                                     #
-# Numerical input. All masses in units of GeV. Updated from PDG 2018. #
-#                                                                     #
-#---------------------------------------------------------------------#
+#----------------------------------------------#
+#                                              #
+# Numerical input. All masses in units of GeV. #
+#                                              #
+#----------------------------------------------#
+
+# Nov 14, 2020: Updated from PDG 2020 and most recent lattice results
 
 # The parameter with 'd' in front denotes the corresponding uncertainty
 # (currently not used anywhere in the code)
@@ -29,27 +31,27 @@ class Num_input(object):
         ### Couplings ###
 
         # The strong coupling constant
-        self.input_parameters['asMZ'] = 0.1181
-        self.input_parameters['dasMZ'] = 0.0011
+        self.input_parameters['asMZ'] = 0.1179
+        self.input_parameters['dasMZ'] = 0.0010
 
         # The Fermi constant
         self.input_parameters['GF'] = 1.1663787e-5
         self.input_parameters['dGF'] = 0.0000006e-5
 
         # The inverse of QED alpha @ MZ
-        self.input_parameters['aMZinv'] = 127.955
-        self.input_parameters['daMZinv'] = 0.01
+        self.input_parameters['aMZinv'] = 127.952
+        self.input_parameters['daMZinv'] = 0.009
 
         # The inverse of QED alpha @ mtau
-        self.input_parameters['amtauinv'] = 133.476
+        self.input_parameters['amtauinv'] = 133.472
         self.input_parameters['damtauinv'] = 0.007
 
         # The inverse of QED alpha at low scales
-        self.input_parameters['alowinv'] = 137.035999139
-        self.input_parameters['dalowinv'] = 0.000000031
+        self.input_parameters['alowinv'] = 137.035999084
+        self.input_parameters['dalowinv'] = 0.000000021
 
         # sin-squared of the weak mixing angle (MS-bar)
-        self.input_parameters['sw2_MSbar'] = 0.23122
+        self.input_parameters['sw2_MSbar'] = 0.23121
         self.input_parameters['dsw2_MSbar'] = 0.00004
 
 
@@ -60,8 +62,8 @@ class Num_input(object):
         self.input_parameters['dMz'] = 0.0021
 
         # Higgs boson mass
-        self.input_parameters['Mh'] = 125.18
-        self.input_parameters['dMh'] = 0.16
+        self.input_parameters['Mh'] = 125.10
+        self.input_parameters['dMh'] = 0.14
 
         # W boson mass
         self.input_parameters['Mw'] = 80.379
@@ -121,13 +123,13 @@ class Num_input(object):
         self.input_parameters['dmc_at_mc'] = 0.03
 
         # strange quark mass, MS-bar at 2 GeV
-        self.input_parameters['ms_at_2GeV'] = 0.096
+        self.input_parameters['ms_at_2GeV'] = 0.093
 
         # down quark mass, MS-bar at 2 GeV
-        self.input_parameters['md_at_2GeV'] = 0.0047
+        self.input_parameters['md_at_2GeV'] = 0.00467
 
         # up quark mass, MS-bar at 2 GeV
-        self.input_parameters['mu_at_2GeV'] = 0.0022
+        self.input_parameters['mu_at_2GeV'] = 0.00216
 
         # top quark mass, MS-bar (converted to MSbar QCD and EW, and run to MZ at 1-loop QCD)
         self.input_parameters['mt_at_MZ'] = 182.
@@ -136,12 +138,20 @@ class Num_input(object):
         ### Low-energy constants for chiral EFT ###
 
         # gA
-        self.input_parameters['gA'] = 1.2723
-        self.input_parameters['dgA'] = 0.0023
+        self.input_parameters['gA'] = 1.2756
+        self.input_parameters['dgA'] = 0.0013
+
+        # Delta u + Delta d
+        self.input_parameters['DeltauDeltad'] = 0.440
+        self.input_parameters['dDeltauDeltad'] = 0.044
+
+        # Deltas
+        self.input_parameters['Deltas'] = -0.035
+        self.input_parameters['dDeltas'] = 0.009
 
         # mG
-        self.input_parameters['mG'] = 0.848
-        self.input_parameters['dmG'] = 0.014
+        self.input_parameters['mG'] = 0.836
+        self.input_parameters['dmG'] = 0.017
 
         # sigmaup
         self.input_parameters['sigmaup'] = 17e-3
@@ -160,40 +170,20 @@ class Num_input(object):
         self.input_parameters['dsigmadn'] = 10e-3
 
         # sigmas
-        self.input_parameters['sigmas'] = 41.3e-3
-        self.input_parameters['dsigmas'] = 7.7e-3
-
-        # Deltaup
-        self.input_parameters['Deltaup'] = 0.897
-        self.input_parameters['dDeltaup'] = 0.027
-
-        # Deltadp
-        self.input_parameters['Deltadp'] = -0.376
-        self.input_parameters['dDeltadp'] = 0.027
-
-        # Deltaun
-        self.input_parameters['Deltaun'] = -0.376
-        self.input_parameters['dDeltaun'] = 0.027
-
-        # Deltadn
-        self.input_parameters['Deltadn'] = 0.897
-        self.input_parameters['dDeltadn'] = 0.027
-
-        # Deltas
-        self.input_parameters['Deltas'] = -0.031
-        self.input_parameters['dDeltas'] = 0.005
+        self.input_parameters['sigmas'] = 52.9e-3
+        self.input_parameters['dsigmas'] = 7.0e-3
 
         # B0mu
-        self.input_parameters['B0mu'] = 6.1e-3
+        self.input_parameters['B0mu'] = 5.8e-3
         self.input_parameters['dB0mu'] = 0.5e-3
 
         # B0md
-        self.input_parameters['B0md'] = 13.3e-3
+        self.input_parameters['B0md'] = 12.4e-3
         self.input_parameters['dB0md'] = 0.5e-3
 
         # B0ms
-        self.input_parameters['B0ms'] = 0.268
-        self.input_parameters['dB0ms'] = 0.003
+        self.input_parameters['B0ms'] = 0.249
+        self.input_parameters['dB0ms'] = 0.009
 
         # Nuclear dipole moments #
 
@@ -203,43 +193,23 @@ class Num_input(object):
         # mun
         self.input_parameters['mun'] = -1.913
 
-        # muup
-        self.input_parameters['muup'] = 1.8045
-
-        # mudp
-        self.input_parameters['mudp'] = -1.097
-
-        # mudn
-        self.input_parameters['mudn'] = 1.8045
-
-        # muun
-        self.input_parameters['muun'] = -1.097
-
         # mus
-        self.input_parameters['mus'] = -0.064
-
-        # ap
-        self.input_parameters['ap'] = 1.793
-
-        # an
-        self.input_parameters['an'] = -1.913
-
-        # F2sp
-        self.input_parameters['F2sp'] = -0.064
+        self.input_parameters['mus'] = -0.036
+        self.input_parameters['dmus'] = 0.021
 
         # nuclear tensor charges (at 2 GeV) #
 
         # gTu
-        self.input_parameters['gTu'] = 0.794
-        self.input_parameters['dgTu'] = 0.015
+        self.input_parameters['gTu'] = 0.784
+        self.input_parameters['dgTu'] = 0.030
 
         # gTd
         self.input_parameters['gTd'] = -0.204
-        self.input_parameters['dgTd'] = 0.008
+        self.input_parameters['dgTd'] = 0.015
 
         # gTs
-        self.input_parameters['gTs'] = 3.2e-4
-        self.input_parameters['dgTs'] = 8.6e-4
+        self.input_parameters['gTs'] = 2.7e-2
+        self.input_parameters['dgTs'] = 1.6e-2
 
         # BT10up
         self.input_parameters['BT10up'] = 3.0
@@ -339,6 +309,50 @@ class Num_input(object):
         self.input_parameters['mu_at_MZ'] = mu(self.input_parameters['Mz'],\
                                                self.input_parameters['mb_at_mb'],\
                                                self.input_parameters['mc_at_mc'], 5, 1)
+
+
+        # Deltaup
+        self.input_parameters['Deltaup'] = (   self.input_parameters['gA']\
+                                             + self.input_parameters['DeltauDeltad'])/2
+        self.input_parameters['dDeltaup'] = np.sqrt(   self.input_parameters['dgA']**2\
+                                                     + self.input_parameters['DeltauDeltad']**2)/2
+
+        # Deltadp
+        self.input_parameters['Deltadp'] = ( - self.input_parameters['gA']\
+                                             + self.input_parameters['DeltauDeltad'])/2
+        self.input_parameters['dDeltadp'] = np.sqrt(   self.input_parameters['dgA']**2\
+                                                     + self.input_parameters['DeltauDeltad']**2)/2
+
+        # Deltaun
+        self.input_parameters['Deltaun'] = self.input_parameters['Deltadp']
+        self.input_parameters['dDeltaun'] = self.input_parameters['dDeltadp']
+
+        # Deltadn
+        self.input_parameters['Deltadn'] = self.input_parameters['Deltaup']
+        self.input_parameters['dDeltadn'] = self.input_parameters['dDeltaup']
+
+        # muup
+        self.input_parameters['muup'] = self.input_parameters['mup'] + self.input_parameters['mun']/2
+
+        # mudp
+        self.input_parameters['mudp'] = self.input_parameters['mup'] + 2*self.input_parameters['mun']
+
+        # mudn
+        self.input_parameters['mudn'] = self.input_parameters['muup']
+
+        # muun
+        self.input_parameters['muun'] = self.input_parameters['mudp']
+
+        # ap
+        self.input_parameters['ap'] = self.input_parameters['mup'] - 1.
+
+        # an
+        self.input_parameters['an'] = self.input_parameters['mun']
+
+        # F2sp
+        self.input_parameters['F2sp'] = self.input_parameters['mus']
+
+
 
 
         if my_input_dict is None: